starting dynamic colouration

+++ Common/src/main/java/dk/sdu/mmmi/cbse/common/data/entityparts/MovingPart.java
@@ -46,7 +46,7 @@ public class MovingPart
         this.maxSpeed = maxSpeed;
     }
 
-    public void setSpeed(float speed) {
+    public void setSpeed(double speed) {
         this.acceleration = speed;
         this.maxSpeed = speed;
     }
